requirements.txt
asgiref==3.5.0
Django==3.2.12
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
graphviz==0.19.2
psycopg2==2.9.3
pydotplus==2.0.2
pyparsing==3.0.7
pytz==2018.4
sqlparse==0.4.2
typing_extensions==4.1.1